refactor : DirtyChecking에 관련된 로직 분리
